pyproject.toml
[project]
name = "daqreminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
]
